INTGR-269: refactor/Tidy up CodeSample folder  - Update icons to follow standards (move to their own folder & make them components)  - Move themes to their own folder  - Update icon CSS
