+++ src/components/CodeSamplesContainer/icons/index.ts
@@ -1,0 +1,3 @@
+export * from './IconCopy';
+export * from './IconDarkModeMoon';
+export * from './IconLightModeSun';
